fix: split large unit test for performance
