+++ src/shared/components/GenericList/test/GenericList.test.js
@@ -35,63 +35,6 @@ describe('GenericList', () => {
     },
   ];
 
-  it('Renders with minimal props', async () => {
-    const { getByText } = render(
-      <ThemeProvider>
-        <GenericList
-          title=""
-          entries={[]}
-          headerRenderer={() => []}
-          rowRenderer={() => []}
-        />
-      </ThemeProvider>,
-    );
-    await waitFor(async () => {
-      await act(async () => {
-        expect(getByText(defaultNotFoundText)).toBeInTheDocument();
-      });
-    });
-  });
-
-  it('Renders custom notFoundMessage props', async () => {
-    const notFoundMessage = 'abcd';
-    const { getByText } = render(
-      <ThemeProvider>
-        <GenericList
-          title=""
-          entries={[]}
-          headerRenderer={() => []}
-          rowRenderer={() => []}
-          notFoundMessage={notFoundMessage}
-        />
-      </ThemeProvider>,
-    );
-    await waitFor(async () => {
-      await act(async () => {
-        expect(getByText(notFoundMessage)).toBeInTheDocument();
-      });
-    });
-  });
-
-  it('Renders title', async () => {
-    const title = 'title';
-    const { getByText } = render(
-      <ThemeProvider>
-        <GenericList
-          title={title}
-          entries={[]}
-          headerRenderer={() => []}
-          rowRenderer={() => []}
-        />
-      </ThemeProvider>,
-    );
-    await waitFor(async () => {
-      await act(async () => {
-        expect(getByText(title)).toBeInTheDocument();
-      });
-    });
-  });
-
   describe('Actions', () => {
     it("Renders actions button when 'actions' prop is provided", async () => {
       const actions = [{ name: 'testaction', handler: () => {} }];
@@ -337,90 +280,6 @@ describe('GenericList', () => {
     await waitFor(async () => {
       await act(async () => {
         expect(getByText(content)).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe('Search', () => {
-    it('Show search field by default', async () => {
-      const { getByRole } = render(
-        <ThemeProvider>
-          <GenericList
-            entries={mockEntries}
-            headerRenderer={mockHeaderRenderer}
-            rowRenderer={mockEntryRenderer}
-          />
-        </ThemeProvider>,
-      );
-      await waitFor(async () => {
-        await act(async () => {
-          expect(getByRole('search')).toBeInTheDocument();
-        });
-      });
-    });
-
-    it("Doesn't show search field when showSearchField is set to false", async () => {
-      const { queryByRole } = render(
-        <ThemeProvider>
-          <GenericList
-            entries={mockEntries}
-            headerRenderer={mockHeaderRenderer}
-            rowRenderer={mockEntryRenderer}
-            searchSettings={{
-              showSearchField: false,
-            }}
-          />
-        </ThemeProvider>,
-      );
-      await waitFor(async () => {
-        await act(async () => {
-          expect(queryByRole('search')).toBeNull();
-        });
-      });
-    });
-
-    it('Shows server error message if dataError prop is true', async () => {
-      const serverErrorMessage = {
-        message: 'Something failed',
-        originalMessage: '',
-      };
-
-      const { queryAllByRole, getByText } = render(
-        <ThemeProvider>
-          <GenericList
-            entries={[]}
-            headerRenderer={mockHeaderRenderer}
-            rowRenderer={mockEntryRenderer}
-            serverDataError={serverErrorMessage}
-          />
-        </ThemeProvider>,
-      );
-
-      await waitFor(async () => {
-        await act(async () => {
-          expect(queryAllByRole('row')).toHaveLength(1);
-          expect(
-            getByText(new RegExp(serverErrorMessage.message, 'i')),
-          ).toBeInTheDocument();
-        });
-      });
-    });
-
-    it('Shows Spinner if dataLoading prop is true', async () => {
-      const { getByLabelText } = render(
-        <ThemeProvider>
-          <GenericList
-            entries={[]}
-            headerRenderer={mockHeaderRenderer}
-            rowRenderer={mockEntryRenderer}
-            serverDataLoading={true}
-          />
-        </ThemeProvider>,
-      );
-      await waitFor(async () => {
-        await act(async () => {
-          expect(getByLabelText('Loading')).toBeInTheDocument();
-        });
       });
     });
   });
